requirements.txt
py-clob-client>=0.34.6
python-dotenv>=1.0.0
rich>=14.0.0
keyboard>=0.13.5
websockets>=12.0
httpx>=0.27.0
